Moved DBMS query methods in their managers
